requirements.txt
Cython>=0.22
cmdstanpy>=1.0.1,!=1.0.2,!=1.0.3
numpy>=1.15.4
pandas>=1.0.4
matplotlib>=2.0.0
LunarCalendar>=0.0.9
convertdate>=2.1.2
holidays>=0.14.2
setuptools>=42
setuptools-git>=1.2
python-dateutil>=2.8.0
tqdm>=4.36.1
wheel>=0.37.0
python-dotenv
pmdarima
scipy
requests